+++ lib/output/sql.go
@@ -47,7 +47,7 @@ func init() {
 			}
 			return NewBatcherFromConfig(conf.SQL.Batching, w, mgr, log, stats)
 		}),
-		Status:  docs.StatusDeprecated,
+		Status:  docs.StatusStable,
 		Batches: true,
 		Async:   true,
 		Version: "3.33.0",
@@ -59,20 +59,7 @@ Runs an SQL prepared query against a target database for each message.`,
 		Description: `
 ## Alternatives
 
-Use the ` + "[`sql_insert`](/docs/components/outputs/sql_insert)" + ` output instead.
-
-## Drivers
-
-The following is a list of supported drivers and their respective DSN formats:
-
-| Driver | Data Source Name Format |
-|---|---|
-` + "| `clickhouse` | [`tcp://[netloc][:port][?param1=value1&...&paramN=valueN]`](https://github.com/ClickHouse/clickhouse-go#dsn)" + `
-` + "| `mysql` | `[username[:password]@][protocol[(address)]]/dbname[?param1=value1&...&paramN=valueN]` |" + `
-` + "| `postgres` | `postgres://[user[:password]@][netloc][:port][/dbname][?param1=value1&...]` |" + `
-` + "| `mssql` | `sqlserver://[user[:password]@][netloc][:port][?database=dbname&param1=value1&...]` |" + `
-
-Please note that the ` + "`postgres`" + ` driver enforces SSL by default, you can override this with the parameter ` + "`sslmode=disable`" + ` if required.`,
+For basic inserts use the ` + "[`sql_insert`](/docs/components/outputs/sql_insert)" + ` output instead. For more complex queries use the ` + "[`sql_raw`](/docs/components/outputs/sql_raw)" + ` output.`,
 		Examples: []docs.AnnotatedExample{
 			{
 				Title: "Table Insert (MySQL)",