+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+## 3.65.0 - TBD
+
+### Added
+
+- New `sql_raw` processor and output.
+
+### Changed
+
+- The `sql` processor and output are no longer marked as deprecated and will therefore not be removed in V4. This change was made in order to provide more time to migrate to the new `sql_raw` processor and output.
+
 ## 3.64.0 - 2022-02-23
 
 ### Added